core tests working
